new way of exposing quasiquotes
Quasiquote interpolators are now immediately available when `scala.meta._` is imported.
However, in order to be usable, quasiquotes require an import from
either `scala.meta.syntactic.quasiquotes._` or `scala.meta.semantic.quasiquotes._`.

In comparison with the previous design, where interpolators had to be
imported from either `scala.meta.syntactic._` or `scala.meta.semantic._`,
this approach is much simpler when both imports have to be active simultaneously.
