--- scalameta/src/main/scala/scala/meta/syntactic/quasiquotes/Macros.scala
@@ -1,24 +1,0 @@
-package scala.meta
-package syntactic
-package quasiquotes
-
-import scala.reflect.macros.whitebox.Context
-import org.scalameta.adt._
-
-class Macros[C <: Context](val c: C) extends AdtReflection with NewLiftables {
-  val u: c.universe.type = c.universe
-  import c.universe.{Tree => _, _}
-  import c.universe.{Tree => ScalaReflectTree}
-  import scala.meta.{Tree => ScalaMetaTree}
-  def apply(macroApplication: ScalaReflectTree, parse: String => ScalaMetaTree): ScalaReflectTree = {
-    val TermQuote = "denied" // TODO: find a cleaner way out of this mess
-    val q"$_($_.apply(..$partlits)).$_.apply[..$_](..$argtrees)($dialect)" = macroApplication
-    val parts = partlits.map{ case q"${part: String}" => part }
-    def ndots(s: String): Int = if (s.endsWith(".")) ndots(s.stripSuffix(".")) + 1 else 0
-    val args = argtrees.zipWithIndex.map{ case (tree, i) => (c.freshName("dummy"), ndots(parts(i)), tree) }
-    val snippet = (parts.init.zip(args).flatMap{ case (part, (id, ndots, _)) => List(if (ndots != 1) part.stripSuffix("." * ndots) else part, s"$id") } :+ parts.last).mkString("")
-    val liveTree: ScalaMetaTree = parse(snippet)
-    val reifiedTree: ScalaReflectTree = implicitly[Liftable[ScalaMetaTree]].apply(liveTree)
-    reifiedTree
-  }
-}
