+++ tests/src/test/scala/ast/InfrastructureSuite.scala
@@ -2,6 +2,7 @@ package scala.meta
 
 import org.scalatest._
 import scala.meta._
+import scala.meta.syntactic.quasiquotes._
 import scala.meta.semantic._
 import scala.meta.syntactic._
 import scala.meta.internal.hygiene.Symbol
